Refactor to have chunks in transactions
