+++ src/main/java/uk/gov/ons/census/action/poller/ChunkPoller.java
@@ -1,0 +1,20 @@
+package uk.gov.ons.census.action.poller;
+
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ChunkPoller {
+  private final ChunkProcessor chunkProcessor;
+
+  public ChunkPoller(ChunkProcessor chunkProcessor) {
+    this.chunkProcessor = chunkProcessor;
+  }
+
+  @Scheduled(fixedDelayString = "${scheduler.frequency}")
+  public void processQueuedCases() {
+    do {
+      chunkProcessor.processChunk();
+    } while (chunkProcessor.isThereWorkToDo()); // Don't go to sleep while there's work to do!
+  }
+}
