+++ src/main/java/uk/gov/ons/census/action/poller/ChunkProcessor.java
@@ -1,0 +1,40 @@
+package uk.gov.ons.census.action.poller;
+
+import java.util.stream.Stream;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+import uk.gov.ons.census.action.model.entity.CaseToProcess;
+import uk.gov.ons.census.action.model.repository.CaseToProcessRepository;
+
+@Component
+public class ChunkProcessor {
+  private final CaseToProcessRepository caseToProcessRepository;
+  private final CaseProcessor caseProcessor;
+
+  @Value("${scheduler.chunksize}")
+  private int chunkSize;
+
+  public ChunkProcessor(
+      CaseToProcessRepository caseToProcessRepository, CaseProcessor caseProcessor) {
+    this.caseToProcessRepository = caseToProcessRepository;
+    this.caseProcessor = caseProcessor;
+  }
+
+  @Transactional(propagation = Propagation.REQUIRES_NEW) // Start a new transaction for every chunk
+  public void processChunk() {
+    try (Stream<CaseToProcess> cases = caseToProcessRepository.findChunkToProcess(chunkSize)) {
+      cases.forEach(
+          caseToProcess -> {
+            caseProcessor.process(caseToProcess);
+            caseToProcessRepository.delete(caseToProcess); // Delete the case from the 'queue'
+          });
+    }
+  }
+
+  @Transactional
+  public boolean isThereWorkToDo() {
+    return caseToProcessRepository.count() > 0;
+  }
+}
